Fixed grunt build process again
