+++ modules/shared/shared.vendor.js
@@ -1,8 +1,5 @@
 /* globals L, proj4, Redux, marked */
 
-import * as Redux from 'redux';
-import marked from 'marked';
-
 (function () {
     'use strict';
